job_hunt_assistant/requirements.txt
crewai>=0.80.0,<1.0
langchain-google-genai>=2.0.0,<3.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.38.0
